Delete the scalar Qwen 3.5 runtime and execute on candle's layers
`qwen35_moe_runtime.rs` was a second implementation of an architecture
candle owns: attention, the gated-delta recurrence, the sparse mixture,
the norms. It existed because for a long time there was nowhere else for
it to live, and its own module docs said it should go once the GPU
pipeline ran. candle #3857 removed the last reason by giving NVFP4 a
packed CPU operator, so a host without a CUDA device no longer chooses
between an eight-times unpack and nothing at all.

Three things came out of the one file, and only one of them was runtime.
The compatibility profile — descriptor, config, layer types, route,
producer and quantization validators — is admission policy, not
execution, and moves to `qwen35_profile.rs` intact. The generation loop
moves onto `ModelForCausalLM`, keeping the parts that were never
architecture-specific: context limits, the wall-clock deadline,
`StopCriteria`, `IncrementalDecoder`, usage accounting, stream control.
The scalar forward pass and its paging working set are gone.

`qwen35_upstream.rs` shrank rather than grew. Its hand-rolled
`residency()`, `projection()`, `WeightResidency` and `Nvfp4Projection`
are one call to candle's `auto_linear`, which reads the packed weight,
the block scales and the tensor scale from the same VarBuilder path and
picks the operator itself.

Three decisions kept on this side rather than delegated:

The `cuda` route is still checked here. Without `nvfp4-cuda` compiled
in, `auto_linear` silently picks the packed CPU operator even on a CUDA
device — the same tokens off the wrong hardware, which is exactly what a
`gpu` route is a claim against. `candle-transformers/nvfp4-cuda` joins
the `candle-cuda` feature in this same change, because splitting the two
is what would leave that trap armed.

`executed_on` gains `cpu_packed_fp4`. The host path is no longer a
fallback: the weight stays at the checkpoint's own footprint.

The `Mutex` is a narrowing, and is documented as one. Upstream's model
owns the KV cache and every linear layer's recurrent state, so requests
to one alias serialize where the scalar runtime kept that state in a
local and did not have to.

690 passing against 715, and the difference is exactly the 25 tests that
lived in the deleted files (32 before across the three, 7 after). The 41
environmental failures are unchanged.

What this gives up: `gated_upstream_layers_agree_with_the_scalar_runtime`
was the proof that these two paths agree, and it goes with the runtime it
compared against. It had never run — the GPU job has no checkpoint — so
the equivalence was never demonstrated, and now cannot be. Nothing in
this change is exercised by CI; it needs a real checkpoint on real
hardware.
